random question patch
