+++ index.html
@@ -8,7 +8,7 @@
     <title>Document</title>
 </head>
 <body>
-    
+
     <script type="module" src="script.js"></script>
 </body>
 </html>
